improve and vectorize code from common
Improve and vectorized some of the function and classes. Scaling is made
into real classes, vectorized and moved to scaling.py. Distances are
vectorized. Size and type checks are implemented for arrays in
check_array() and check_arrays() in common. find_mse and find_accuracy
are also vectorized. Add integer_one_hot_encode.